FastApi_Introduction/my-movie-api/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
h11==0.14.0
idna==3.4
pydantic==1.10.6
PyJWT==2.6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
